# docs/api/openapi-spec.yaml
openapi: 3.1.0
info:
  title: Nexary API
  description: |
    Nexary is a full-stack AI platform featuring multi-provider AI chat, RAG (Retrieval-Augmented Generation),
    team-based workspace management, and enterprise-grade compliance features.

    ## Authentication
    Most endpoints require authentication via Stack Auth. Include the session token in cookies.
    Some endpoints also support API key authentication via the `Authorization` header.

    ## Rate Limiting
    API requests are rate-limited based on subscription plan.
  version: 1.0.0
  contact:
    name: Nexary API Support
    email: support@nexary.ai
  license:
    name: Proprietary
servers:
  - url: https://api.nexary.ai
    description: Production server
  - url: https://staging-api.nexary.ai
    description: Staging server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/rag/packages/{id}/documents/bulk/delete:
    post:
      tags: [RAG]
      summary: Bulk delete documents
      description: Soft-delete multiple documents
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/RAGPackageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [documentIds]
              properties:
                documentIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Documents deleted
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List all users
      description: Admin endpoint to list all users (requires admin role)
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: Stack Auth session cookie
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API key in format `Bearer {api_key}`
  parameters:
...
    RAGPackageId:
      name: id
      in: path
      required: true
      description: RAG package ID
      schema:
        type: string
        format: uuid
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
...
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...